revert ecosystem page path
